Rebind descriptors set directly if possible.
Avoids going through a full hashing song and dance if we just bumped the
UBO block offset.

To make this path more likely, we bump the UBO size to
VULKAN_MAX_UBO_SIZE if the block
allocator has enough room for it, so we don't have to encode a different
range (i.e., a new descriptor set).