Refactor the memory allocator a bit.
Use a "global" object pool instead of per-class allocator, which is a
bit inefficient.

Needed to refactor the locking scheme to be allocator global instead.
Made more sense to move the locking responsibility to caller.
For back-to-back free operations and multi-allocation scenarios it makes
mores sense to lock once anyways ...

Make some functions private to ensure that the memory allocator is only
called from where we expect them.